final restructring
