+++ graphql/resolvers/booking.js
@@ -1,0 +1,40 @@
+const Booking = require("../../models/booking");
+const Event = require("../../models/event");
+const { transformEvent, transformBooking } = require("./merge");
+
+module.exports = {
+  bookings: async () => {
+    try {
+      const bookings = await Booking.find();
+      return bookings.map((booking) => {
+        return transformBooking(booking);
+      });
+    } catch (error) {
+      throw error;
+    }
+  },
+  bookEvent: async (args) => {
+    try {
+      const fetchedEvent = await Event.findOne({ _id: args.eventId });
+      const booking = new Booking({
+        user: "65103e2690e6049565a26cba",
+        event: fetchedEvent,
+      });
+      const result = await booking.save();
+      return transformBooking(result);
+    } catch (error) {
+      throw error;
+    }
+  },
+  cancelBooking: async (args) => {
+    try {
+      const booking = await Booking.findById(args.bookingId).populate("event");
+      const event = transformEvent(booking.event);
+      console.log(event);
+      await Booking.deleteOne({ _id: args.bookingId });
+      return event;
+    } catch (error) {
+      throw error;
+    }
+  },
+};
