+++ graphql/resolvers/events.js
@@ -1,0 +1,40 @@
+const Event = require("../../models/event");
+const User = require("../../models/user");
+const { transformEvent } = require("./merge");
+
+module.exports = {
+  events: async () => {
+    try {
+      const events = await Event.find();
+      return events.map((event) => {
+        return transformEvent(event);
+      });
+    } catch (error) {
+      throw err;
+    }
+  },
+  createEvent: async (args) => {
+    const event = new Event({
+      title: args.eventInput.title,
+      description: args.eventInput.description,
+      price: +args.eventInput.price,
+      date: new Date(args.eventInput.date),
+      creator: "65103e2690e6049565a26cba", // just hardcoded for now
+    });
+    let createdEvent;
+
+    try {
+      const savedEvent = await event.save();
+      createdEvent = transformEvent(savedEvent);
+      const creator = await User.findById("65103e2690e6049565a26cba");
+      if (!creator) {
+        throw new Error("User not found.");
+      }
+      creator.createdEvents.push(event);
+      await creator.save();
+      return createdEvent;
+    } catch (error) {
+      throw error;
+    }
+  },
+};
