+++ graphql/resolvers/index.js
@@ -1,165 +1,11 @@
-const Event = require("../../models/event");
-const User = require("../../models/user");
-const Booking = require("../../models/booking");
-const bcrypt = require("bcryptjs");
+const authResolver = require("./auth");
+const eventsResolver = require("./events");
+const bookingResolver = require("./booking");
 
-const events = async (eventIds) => {
-  try {
-    const events = await Event.find({ _id: { $in: eventIds } });
-    events.map((event) => {
-      return {
-        ...event._doc,
-        _id: event.id,
-        date: new Date(event._doc.date).toISOString(),
-        creator: user.bind(this, event.creator),
-      };
-    });
-    return events;
-  } catch (error) {
-    throw error;
-  }
+const rootResolver = {
+  ...authResolver,
+  ...eventsResolver,
+  ...bookingResolver,
 };
 
-const user = async (userId) => {
-  try {
-    const user = await User.findById(userId);
-    return {
-      ...user._doc,
-      _id: user.id,
-      createdEvents: events.bind(this, user._doc.createdEvents),
-    };
-  } catch (error) {
-    throw error;
-  }
-};
-
-const singleEvent = async (eventId) => {
-  try {
-    const event = await Event.findById(eventId);
-    return {
-      ...event._doc,
-      _id: event.id,
-      creator: user.bind(this, event.creator),
-    };
-  } catch (error) {
-    throw error;
-  }
-};
-
-module.exports = {
-  events: async () => {
-    try {
-      const events = await Event.find();
-      return events.map((event) => {
-        return {
-          ...event._doc,
-          _id: event.id,
-          date: new Date(event._doc.date).toISOString(),
-          creator: user.bind(this, event._doc.creator),
-        };
-      });
-    } catch (error) {
-      throw err;
-    }
-  },
-  bookings: async () => {
-    try {
-      const bookings = await Booking.find();
-      return bookings.map((booking) => {
-        return {
-          ...booking.event,
-          _id: booking._id,
-          user: user.bind(this, booking._doc.user),
-          event: singleEvent.bind(this, booking._doc.event),
-          createdAt: new Date(booking._doc.createdAt).toISOString(),
-          updatedAt: new Date(booking._doc.updatedAt).toISOString(),
-        };
-      });
-    } catch (error) {
-      throw error;
-    }
-  },
-  createEvent: async (args) => {
-    const event = new Event({
-      title: args.eventInput.title,
-      description: args.eventInput.description,
-      price: +args.eventInput.price,
-      date: new Date(args.eventInput.date),
-      creator: "65103e2690e6049565a26cba", // just hardcoded for now
-    });
-    let createdEvent;
-
-    try {
-      const savedEvent = await event.save();
-      createdEvent = {
-        ...savedEvent._doc,
-        _id: savedEvent._doc._id.toString(),
-        date: new Date(savedEvent._doc.date).toISOString(),
-        creator: user.bind(this, savedEvent._doc.creator),
-      };
-      const creator = await User.findById("65103e2690e6049565a26cba");
-      if (!creator) {
-        throw new Error("User not found.");
-      }
-      creator.createdEvents.push(event);
-      await creator.save();
-      return createdEvent;
-    } catch (error) {
-      throw error;
-    }
-  },
-  createUser: async (args) => {
-    try {
-      const existingUser = await User.findOne({ email: args.userInput.email });
-      if (existingUser) {
-        throw new Error("User exist already.");
-      }
-      const hashedPassword = await bcrypt.hash(args.userInput.password, 12);
-
-      const user = new User({
-        email: args.userInput.email,
-        password: hashedPassword,
-      });
-
-      const result = await user.save();
-      return { ...result._doc, password: null, _id: result.id };
-    } catch (error) {
-      throw err;
-    }
-  },
-  bookEvent: async (args) => {
-    try {
-      const fetchedEvent = await Event.findOne({ _id: args.eventId });
-      const booking = new Booking({
-        user: "65103e2690e6049565a26cba",
-        event: fetchedEvent,
-      });
-      const result = await booking.save();
-      return {
-        ...result._doc,
-        _id: result.id,
-        createdAt: new Date(result._doc.createdAt).toISOString(),
-        updatedAt: new Date(result._doc.updatedAt).toISOString(),
-        user: user.bind(this, result._doc.user),
-        event: singleEvent.bind(this, result._doc.event),
-      };
-    } catch (error) {
-      throw error;
-    }
-  },
-  cancelBooking: async (args) => {
-    try {
-      const booking = await Booking.findById(args.bookingId).populate("event");
-      const event = {
-        ...booking.event._doc,
-        _id: booking.id,
-        creator: user.bind(this, booking.event._doc.creator),
-      };
-      console.log(event);
-      await Booking.deleteOne({ _id: args.bookingId });
-      return event;
-    } catch (error) {
-      throw error;
-    }
-  },
-};
+module.exports = rootResolver;
